+++ sql/queries/feed_follows.sql
@@ -10,4 +10,4 @@ SELECT * FROM feed_follows WHERE id = $1;
 SELECT * FROM feed_follows WHERE user_id = $1;
 
 -- name: DeleteFeedFollowsByID :exec
-DELETE FROM feed_follows WHERE id = $1; 
+DELETE FROM feed_follows WHERE id = $1 AND user_id = $2;
